fix: update ON10 validity range
